requirements.txt
disnake==2.7.0
emoji==2.7.0
mcstatus==11.0.0
python_a2s==1.3.0
PyYAML==6.0.1
